Ver 1.4
Ver 1.4 uses a queue instead of a stack, so now it finds all possible paths and still solves the maze. Input, output, etc remains the same, just runs differently under the hood. The stack code is still there, just commented out so it doesn't run, so just leave it that way and it will run off the queue code instead of the stack code.