Ver 1.7
Ver 1.7 uses the maze to graph conversion from Ver 1.6, but then uses a Breadth First Search instead to reach the end of the maze. Output is the same as Ver 1.4. Note about provided output file, all spaces are filled in because of the specific case of the maze from the input file. Mathmatically speaking, every space would have been visited before reaching the end using a breadth first search. This is not true for every maze, the code will find the shortest route given any maze, but in this particular case, visited all squares in it's search.